HTTP responses and data vlidation on all layers

+++ StockExchange.DAL/DataModel/EodPrice.cs
@@ -10,7 +10,7 @@
         [Required]
         public DateTime Date { get; set; } = DateTime.Now;
         [Required]
-        public float ClosePrice { get; set; }
+        public decimal ClosePrice { get; set; }
         [Required]
         public StockSymbol? stockSymbol { get; set; } // n-1
         //foreign key
