+++ StockExchange.Domain.Model/EodPriceModel.cs
@@ -7,7 +7,7 @@
     {
         public int ID { get; set; }
         public DateTime Date { get; set; }
-        public float ClosePrice { get; set; }
+        public decimal ClosePrice { get; set; }
         public int StockSymbolId { get; set; }
         [JsonIgnore]
